Mount /incidents client-only via dynamic({ ssr: false })
Same shape as ADR-062 §1 for AppShell: extract the body of
incidents/page.tsx into IncidentsClient.tsx and rewrite page.tsx as a
thin dynamic boundary. The browser-API reads that used to live in a
useEffect (URL → localStorage → 'default' fallback) now run inside a
useState lazy initializer, so the resolved project lands on the first
render instead of after a re-render.

That collapses the double-fetch the deferred shape produced: the
useEffect([project]) consumer fires once against the resolved project
rather than first against 'default' and then against the real value.
The typeof window guard stays in the initializer as belt-and-suspenders
— if someone later removes { ssr: false } the page degrades to a
flicker, not a crash.
